==> notebooks/01-data-store.ipynb <==
import requests
import json
from mlProject.utils.ura_token import ura_token
from pandas import json_normalize
import pandas as pd
import psycopg2
import uuid
import pandas.io.sql as psql


from datetime import datetime
from typing import List, Optional
#from pydantic import BaseModel
import json
import psycopg2.extras

# call it in any place of your program
# before working with UUID objects in PostgreSQL
psycopg2.extras.register_uuid()

from dotenv import load_dotenv
import os

# Load environment variables from the .env file (if present)
load_dotenv()

# Access environment variables as if they came from the actual environment
ura_access_key = os.getenv('ura_access_key')
ura_access_key

u=ura_token(ura_access_key)

df = pd.DataFrame()
for batch in range(1,5):
    url=f"https://www.ura.gov.sg/uraDataService/invokeUraDS?service=PMI_Resi_Transaction&batch={batch}"
    myobj = {'AccessKey': u.get_accesskey(), 'token': u.get_token(), 'User-Agent': 'Mozilla/5.0'}
    responses = requests.post(url, headers=myobj)

    results = responses.json()['Result']
    flattendata = json_normalize(results,'transaction',['project','street','marketSegment','x','y'],errors='ignore')
    flattendata['contractDate'] = pd.to_datetime(flattendata['contractDate'],format='%m%y')
    df = pd.concat([df,flattendata])
df.head()
    

df.shape

now_str = datetime.now().strftime("%Y%m%d%H%M%S")
df = df.reset_index().rename(columns={'index':'id'})
df['id'] = df['id'].apply(lambda x: str(x)+'_'+now_str)
df.head(2)

df.to_csv('transaction.csv')

df['noOfUnits'] = df['noOfUnits'].astype(int)
numerical_cols = ['area','price','nettPrice','x','y']
for col in numerical_cols:
    df[col] = df[col].astype(float)

df.head()

df.dtypes

from pandantic import BaseModel
from pydantic.types import StrictInt,StrictFloat,StrictStr


class DataFrameSchema(BaseModel):
    area: float
    floorRange: str 
    noOfUnits: StrictInt
    contractDate: Optional[datetime] 
    typeOfSale: str
    price: float
    propertyType: str
    district: str
    typeOfArea: str
    tenure: str
    nettPrice: float
    project: str
    street: str
    marketSegment: str
    x: float
    y: float  




df_valid = DataFrameSchema.parse_df(
    dataframe=df,
    errors="filter",
)

df[df.duplicated(subset=['id'])]

def save_dataframe_to_postgresql(df: pd.DataFrame, table_name: str, conn_params: dict):
    connection = psycopg2.connect(**conn_params)
    cursor = connection.cursor()
    
    # Create the table if it doesn't exist
    cursor.execute(f"""
        CREATE TABLE IF NOT EXISTS {table_name} (
            id TEXT PRIMARY KEY,
            area FLOAT,
            floor_range TEXT,
            no_of_units INT,
            contract_date TIMESTAMP,
            type_of_sale TEXT,
            price FLOAT,
            property_type TEXT,
            district TEXT,
            type_of_area TEXT,
            tenure TEXT,
            nett_price FLOAT,
            project TEXT,
            street TEXT,
            market_segment TEXT,
            x FLOAT,
            y FLOAT
        );
    """)
    connection.commit()
    
    # Insert data from DataFrame to PostgreSQL
    for index, row in df.iterrows():
        insert_query = f'''
        INSERT INTO {table_name} (id, area, floor_range, no_of_units, contract_date, type_of_sale, price, property_type, district, type_of_area, tenure, nett_price, project, street, market_segment, x, y)
        VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
        '''
        cursor.execute(insert_query, (
            str(uuid.uuid4()), row['area'], row['floorRange'], row['noOfUnits'], row['contractDate'], 
            row['typeOfSale'], row['price'], row['propertyType'], row['district'], row['typeOfArea'], 
            row['tenure'], row['nettPrice'], row['project'], row['street'], row['marketSegment'], 
            row['x'], row['y']
        ))

    # Commit the changes
    connection.commit()
    cursor.close()
    connection.close()



def get_dataframe_from_sql(table_name: str, conn_params: dict):
    connection = psycopg2.connect(**conn_params)
    cursor = connection.cursor()
    df = psql.read_sql(f"Select * from {table_name}", connection)
    return df

# Connect to PostgreSQL
conn_params = {
    "host":"localhost",
    "database":"mydatabase",
    "user":"myuser",
    "password":"mypassword"

}

INGESTION=False
table_name = "propertypricetable"
if INGESTION:
    save_dataframe_to_postgresql(df, table_name,conn_params)

df_fetch = get_dataframe_from_sql(table_name, conn_params)
df_fetch.head()

connection = psycopg2.connect(**conn_params)
cursor = connection.cursor()
cursor.execute("""SELECT table_name FROM information_schema.tables
       WHERE table_schema = 'public'""")
for table in cursor.fetchall():
    print(table)
cursor.close()
connection.close()

df_ec = df.loc[(df['propertyType']=='Executive Condominium')&(df['typeOfSale']=='3')].sort_values(by=['contractDate'])

df_ec.head(2)

df_ec.tail(2)

df_ec['noOfUnits'].unique()

df_ec['typeOfArea'].unique()

df_ec['tenure'].unique()

df_ec['tenure_start'] = df_ec['tenure'].apply(lambda x: int(x.split(' ')[-1]))
print(df_ec['tenure_start'].unique())
df_ec['building_age'] = df_ec['contractDate'].dt.year - df_ec['tenure_start']
df_ec.head()

#
