==> notebooks/05-test-api.ipynb <==
```bash
source .venv/bin/activate
python3 -m uvicorn main:app --reload
```

import requests
import pandas as pd 


df_test = pd.read_csv('../data/processed/test.csv')
data = df_test.drop(columns=['price']).iloc[0,:].to_dict()

data

response = requests.post(url="http://54.255.65.238:8000/predict_price", params=data)
print(response.json())
